requirements.txt
# Core LangChain packages
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0

# LangSmith for tracing and monitoring
langsmith>=0.1.99

# Popular LLM integrations
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0

# Vector stores
langchain-chroma>=0.1.4
langchain-pinecone>=0.2.0
faiss-cpu>=1.8.0

# Document loaders and text processing
pypdf2>=3.0.1
python-docx>=1.1.2
beautifulsoup4>=4.12.3
tiktoken>=0.7.0
nltk>=3.8.1

# Web scraping and APIs
requests>=2.32.3
aiohttp>=3.10.5

# Development and testing tools
pytest>=8.3.2
pytest-asyncio>=0.24.0
black>=24.8.0
isort>=5.13.2
mypy>=1.11.2
ruff>=0.6.4
pre-commit>=3.8.0

# Jupyter for notebooks
jupyter>=1.1.1
jupyterlab>=4.2.5
ipykernel>=6.29.5

# Environment management
python-dotenv>=1.0.1

# Additional useful packages
pydantic>=2.8.2
tenacity>=8.5.0
sqlalchemy>=2.0.32
redis>=5.0.8
celery>=5.4.0